docs: improve documentation
