+++ plugins/receivers/slack/option.go
@@ -6,6 +6,7 @@ import (
 
 type ClientOption func(*Client)
 
+// ClientWithHTTPClient assigns custom http client when creating a slack client
 func ClientWithHTTPClient(httpClient *http.Client) ClientOption {
 	return func(c *Client) {
 		c.httpClient = httpClient
@@ -14,12 +15,14 @@ func ClientWithHTTPClient(httpClient *http.Client) ClientOption {
 
 type ClientCallOption func(*clientData)
 
+// CallWithGoSlackClient uses a custom slack client when calling slack API
 func CallWithGoSlackClient(gsc GoSlackCaller) ClientCallOption {
 	return func(c *clientData) {
 		c.goslackClient = gsc
 	}
 }
 
+// CallWithClientSecret uses a new client with client ID, secret, and auth code to call slack API
 func CallWithClientSecret(authCode string, clientID, clientSecret string) ClientCallOption {
 	return func(c *clientData) {
 		c.authCode = authCode
@@ -28,6 +31,7 @@ func CallWithClientSecret(authCode string, clientID, clientSecret string) Client
 	}
 }
 
+// CallWithToken uses access token to call slack API
 func CallWithToken(token string) ClientCallOption {
 	return func(c *clientData) {
 		c.token = token
